Added a CreateNewGame component
