+++ app/javascript/components/CreateNewGame/Location.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import Box from '@material-ui/core/Box';
 import { makeStyles } from '@material-ui/core/styles';
 import TextField from '@material-ui/core/TextField';
-import GameInfoStyles from './GameInfoStyles';
+import CreateNewGameStyles from './CreateNewGameStyles';
 
-const useStyles = makeStyles(GameInfoStyles, (theme) => ({
+const useStyles = makeStyles(CreateNewGameStyles, (theme) => ({
   textField: {
     marginLeft: theme.spacing(1),
     marginRight: theme.spacing(1),
@@ -16,15 +16,13 @@ const Location = () => {
   const classes = useStyles();
 
   return (
-    <form className={classes.root} noValidate autoComplete="off">
-      <Box>
-        <TextField
-          className={classes.locationField}
-          id="standard-basic"
-          label="Location"
-        />
-      </Box>
-    </form>
+    <Box>
+      <TextField
+        className={classes.locationField}
+        id="standard-basic"
+        label="Location"
+      />
+    </Box>
   );
 };
