+++ app/javascript/components/CreateNewGame/index.jsx
@@ -1,0 +1,3 @@
+import CreateNewGame from "./CreateNewGame";
+
+export default CreateNewGame;
