+++ app/javascript/components/GameDetails/GameDetails.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ScoreBoard from './ScoreBoard';
 import GameInfo from './GameInfo';
